Added trend arrows and made UI more user-friendly
Rate of change is now calculated between the last three available points
and an arrow and brief description is displayed. The arrows suck because
I can't use unicode characters in the terminal, but if I ever get a GUI
for this, I'll use fancy unicode arrows.

Also, data in the UI is displayed in morw is displayed in more
appropiate units (g instead of mg) and is rounded to not take up an
uneccessary amount of space. The actual variables still retain the data
in its long form, though, so quality of teh data will not be effected.