Fixed graphs + fixing update system
Fixed some of the display errors with the graphs, now I'm just trying to
find a better way than having the user close and then system re-open the
graph each time the display updates. I want the graph to clear itself
and refresh automatically, but I haven't found a awy to do that yet.